requirements.txt
aiodns==3.1.1
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
argcomplete==3.1.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
beautifulsoup4==4.12.3
blinker==1.9.0
Brlapi==0.8.5
ccsm==0.9.14.2
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==25.1.2
colorama==0.4.6
command-not-found==0.3
compizconfig-python==0.9.14.2
configobj==5.0.8
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
discord.py==2.3.2
distro==1.9.0
distro-info==1.7+build1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
evdev==1.7.0
Flask==3.1.1
frozenlist==1.4.0
gyp==0.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httplib2==0.20.4
httpx==0.26.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.4
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
openai==1.12.0
packaging==24.0
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.3
platformdirs==4.2.0
protobuf==4.21.12
psutil==5.9.8
ptyprocess==0.7.0
pycairo==1.25.1
pycares==4.4.0
pycups==2.0.1
pydantic==1.10.14
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
setproctitle==1.3.3
setuptools==68.1.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
systemd-python==235
texttable==1.6.7
tqdm==0.0.0
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
userpath==1.9.1
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.1.3
wheel==0.42.0
xdg==5
xkit==0.0.0
yarl==1.8.2
